+++ components/shared/RatingBoard/styles.js
@@ -5,7 +5,6 @@ const { height: windowHeight } = Dimensions.get("window");
 const styles = StyleSheet.create({
   container: {
     flexDirection: "row",
-
     height: windowHeight * 0.05,
   },
 });
